Use upstream clojure mode - close #26
Also verified #69 and
https://github.com/LightTable/LightTable/issues/2100 are fixed
